requirements.txt
flask
pillow
ultimateAlprSdk